+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 
 ### ✨ Added
 
+- **Respect `replace_path` in the mod descriptor file.**
+
 - **Order of Battle (`history/units/*.txt`) scanner.** Self-validates division templates.
 
 - **Unit (`common/units/*.txt`) scanner.** Units are validated within OOBs.
@@ -15,6 +17,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 ### 🩹 Fixed
 
 - **Fixed an air wings OOB edge-case where HoM was emitting a `duplicate_key` warning for multiple `name` keys being under the same state scope under respective equipment types (planes).** This is *required* to name air wings when there are more than one of them in the same state/airfield.
+
+- **Fixed an issue where `did_change_watched_files` only updated the scan data, never re-validated open documents whose diagnostics could depend on that data.**
 
 ## [v0.17.0] - 2026-06-07